# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "Coldcat backend API", "version": "0.1.0", "description": "Content, observation and disk-management API. Imports run offline with the server stopped. GETs are read-only. IDs, counts and bytes are decimal strings. Only complete inventories are visible."},
  "servers": [{"url": "http://127.0.0.1:8080"}],
  "paths": {
...
    "/api/v1/contents": {
      "get": {
        "summary": "List distinct contents with current redundancy filters",
        "description": "Ordered by content ID ascending. Disk/directory filters select membership; counts remain catalog-wide. Current scope includes only currently present content. History includes contents from all complete inventories. Replica bounds require current scope; exact bounds cannot be combined with range bounds. Keep normalized filters and limit unchanged when following a cursor. Successful offline imports invalidate cursors; unchanged restarts, failed imports and recovery preserve them. Unknown, repeated and empty parameters are rejected except directory, whose empty value means root. Missing disks return 404; missing directories return an empty page.",
        "parameters": [
          {"$ref": "#/components/parameters/Scope"},
          {"name": "disk_id", "in": "query", "schema": {"$ref": "#/components/schemas/ID"}},
          {"name": "directory", "in": "query", "description": "Requires disk_id. Case-sensitive disk-relative directory, with / separators, no leading/trailing slash or . or .. segments. Maximum 1024 UTF-8 bytes. Descendant segment boundaries and literal wildcard characters are preserved. Empty means root.", "schema": {"type": "string"}, "allowEmptyValue": true},
          {"name": "replica_metric", "in": "query", "schema": {"type": "string", "enum": ["disks", "locations"], "default": "disks"}},
          {"name": "other_replicas", "in": "query", "description": "Exact current count minus one; cannot combine with min/max bounds.", "schema": {"$ref": "#/components/schemas/Decimal"}},
          {"name": "min_other_replicas", "in": "query", "description": "Inclusive minimum current count minus one.", "schema": {"$ref": "#/components/schemas/Decimal"}},
          {"name": "max_other_replicas", "in": "query", "description": "Inclusive maximum current count minus one; must be at least the minimum.", "schema": {"$ref": "#/components/schemas/Decimal"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 200, "default": 50}},
          {"name": "cursor", "in": "query", "schema": {"type": "string", "minLength": 1, "maxLength": 16384}, "description": "Opaque next_cursor from the preceding page"}
        ],
        "responses": {
          "200": {"description": "Content page with applied filters and catalog revision", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentPage"}, "examples": {
            "twoOtherDisks": {"value": {"scope": "current", "revision": "3", "filters": {"disk_id": null, "directory": "", "replica_metric": "disks", "other_replicas": "2", "min_other_replicas": null, "max_other_replicas": null}, "items": [{"id": "1", "hash": {"algorithm": "sha256", "hex": "ab"}, "size": "0", "scope": "current", "location_count": "4", "disk_count": "3", "observation_count": "4", "current_location_count": "4", "current_disk_count": "3"}], "next_cursor": null}},
            "empty": {"value": {"scope": "current", "revision": "0", "filters": {"disk_id": null, "directory": "", "replica_metric": "disks", "other_replicas": null, "min_other_replicas": null, "max_other_replicas": null}, "items": [], "next_cursor": null}}
          }}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "405": {"description": "Only GET and HEAD are allowed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "409": {"description": "Inventory revision changed; restart pagination", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "stale_cursor", "message": "cursor belongs to a different inventory revision"}}}}},
          "500": {"$ref": "#/components/responses/Internal"},
          "503": {"$ref": "#/components/responses/Unavailable"}
        }
      }
    },
    "/api/v1/contents/lookup": {
      "get": {
        "summary": "Look up an exact algorithm/hash identity",
        "description": "Different algorithms never establish equivalence. Hex is case-insensitive on input, lowercase on output. Accepts the hash lengths currently accepted by the importer.",
        "parameters": [
          {"name": "hash_type", "in": "query", "required": true, "schema": {"$ref": "#/components/schemas/Algorithm"}, "example": "sha256"},
          {"name": "hash", "in": "query", "required": true, "schema": {"type": "string", "pattern": "^(?:[0-9a-fA-F]{2})+$"}, "example": "AB"},
          {"$ref": "#/components/parameters/Scope"}
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/Content"},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "503": {"$ref": "#/components/responses/Unavailable"},
          "500": {"$ref": "#/components/responses/Internal"}
        }
      }
    },
    "/api/v1/contents/{id}": {
      "get": {
        "summary": "Read content and scoped replica counts",
        "parameters": [{"$ref": "#/components/parameters/ID"}, {"$ref": "#/components/parameters/Scope"}],
        "responses": {
          "200": {"$ref": "#/components/responses/Content"},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "503": {"$ref": "#/components/responses/Unavailable"},
          "500": {"$ref": "#/components/responses/Internal"}
        }
      }
    },
    "/api/v1/contents/{id}/observations": {
      "get": {
        "summary": "Page through current locations or historical observations",
        "description": "Ordered by observation ID ascending. Current uses the latest complete inventory of each disk by (captured_at,id). History includes repeated observations; location_count instead deduplicates disk/path pairs. Keep content ID, scope and limit unchanged when following a cursor. Revision is the highest completed snapshot ID, or 0. Successful publication invalidates cursors; restart, empty-disk creation and failed-import cleanup preserve them. Cursors are for the issuing catalog. Unknown, empty and repeated scalar parameters are rejected.",
        "parameters": [
          {"$ref": "#/components/parameters/ID"},
          {"$ref": "#/components/parameters/Scope"},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 200, "default": 50}},
          {"name": "cursor", "in": "query", "schema": {"type": "string", "minLength": 1, "maxLength": 2048}, "description": "Opaque next_cursor from the preceding page"}
        ],
        "responses": {
          "200": {"description": "Observation page; historical-only content has an empty current list", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ObservationPage"}, "examples": {
            "lastPage": {"value": {"scope": "current", "revision": "3", "items": [{
              "observation": {"id": "5", "content_id": "1", "snapshot_id": "3", "path": "zéro/a", "mtime": null},
              "snapshot": {"id": "3", "disk_id": "3", "state": "complete", "captured_at": "1970-01-01T00:00:10.000000123Z", "imported_at": "2026-10-05T12:00:00Z", "capture_provenance": "explicit", "input_path": "/inventories/disk-2.cshd", "input_format": "cshd", "file_count": "1", "content_count": "1"},
              "disk": {"id": "3", "label": "disk-2", "serial": "", "capacity": "100"}, "size": "0", "is_current": true
            }], "next_cursor": null}},
            "historicalOnly": {"value": {"scope": "current", "revision": "3", "items": [], "next_cursor": null}}
          }}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"description": "Inventory changed since the cursor was issued; restart pagination", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "stale_cursor", "message": "cursor belongs to a different inventory revision"}}}}},
          "503": {"$ref": "#/components/responses/Unavailable"},
          "500": {"$ref": "#/components/responses/Internal"}
        }
      }
    },
    "/api/v1/observations/{id}": {
      "get": {
        "summary": "Read an observation, inventory and disk with other current replica counts",
        "description": "Other counts exclude the observation's disk/path or disk only where currently present for that content. Historical-only contents have zero other replicas. Embedded content uses current scope.",
        "parameters": [{"$ref": "#/components/parameters/ID"}],
        "responses": {
          "200": {"description": "Observation detail", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ObservationSummary"}}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "503": {"$ref": "#/components/responses/Unavailable"},
          "500": {"$ref": "#/components/responses/Internal"}
        }
      }
    },
    "/api/v1/disks": {
      "get": {
        "summary": "List disk metadata and latest complete inventories",
        "description": "Disk ID ascending. Keep limit unchanged when following a cursor. The first page fixes the maximum disk ID; disks created later are excluded from that traversal. Metadata edits appear as currently stored on each page and do not invalidate cursors. Successful offline imports invalidate cursors; unchanged restarts and failed imports preserve them. This is not a historical metadata snapshot. Unknown, repeated and empty parameters are rejected.",
        "parameters": [
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 200, "default": 50}},
          {"name": "cursor", "in": "query", "schema": {"type": "string", "minLength": 1, "maxLength": 2048}}
        ],
        "responses": {
          "200": {"description": "Disk page", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DiskPage"}, "example": {"revision": "0", "items": [{"id": "1", "label": "archive", "notes": null, "serial": null, "capacity": "1000000000000", "latest_snapshot": null}], "next_cursor": null}}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "405": {"$ref": "#/components/responses/DiskCollectionMethod"},
          "409": {"$ref": "#/components/responses/StaleCursor"},
          "500": {"$ref": "#/components/responses/Internal"},
          "503": {"$ref": "#/components/responses/Unavailable"}
        }
      },
      "post": {
        "summary": "Create a disk",
        "description": "Requires a unique case-sensitive nonempty label and capacity as a decimal string in the range 0..9223372036854775807. Optional notes/serial may be omitted, null or empty. Unknown fields, trailing JSON values and query parameters are rejected. Body maximum 65536 bytes; Content-Type must be application/json.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateDisk"}, "example": {"label": "archive", "capacity": "1000000000000", "notes": "offline backup"}}}},
        "responses": {
          "201": {"description": "Created disk", "headers": {"Location": {"description": "Created disk detail route", "schema": {"type": "string"}, "example": "/api/v1/disks/1"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DiskDetail"}, "example": {"id": "1", "label": "archive", "notes": "offline backup", "serial": null, "capacity": "1000000000000", "latest_snapshot": null, "cataloged": null}}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "405": {"$ref": "#/components/responses/DiskCollectionMethod"},
          "409": {"$ref": "#/components/responses/DiskConflict"},
          "413": {"$ref": "#/components/responses/TooLarge"},
          "415": {"$ref": "#/components/responses/UnsupportedMedia"},
          "500": {"$ref": "#/components/responses/Internal"},
          "503": {"$ref": "#/components/responses/Unavailable"}
        }
      }
    },
    "/api/v1/disks/{id}": {
      "get": {
        "summary": "Read disk metadata and cataloged size completeness",
        "description": "Latest complete inventory by (captured_at,id). Cataloged bytes sum known content sizes per file path, including repeated contents. Unknown sizes are counted separately. These are inventory subtotals, not measured disk usage/free space. A disk without an inventory has null latest_snapshot and cataloged; a complete empty inventory has zero counts and size_complete=true. Size-sum overflow is an explicit server error. No query parameters are accepted.",
        "parameters": [{"$ref": "#/components/parameters/ID"}],
        "responses": {
          "200": {"description": "Disk detail", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DiskDetail"}, "example": {"id": "1", "label": "archive", "notes": null, "serial": "ABC", "capacity": "1000000000000", "latest_snapshot": {"id": "3", "disk_id": "1", "state": "complete", "captured_at": "2023-01-01T00:00:00Z", "imported_at": "2026-10-06T12:00:00Z", "capture_provenance": "explicit", "input_path": "/inventories/archive.cshd", "input_format": "cshd", "file_count": "4", "content_count": "3"}, "cataloged": {"file_count": "4", "content_count": "3", "known_bytes": "20", "unknown_size_file_count": "1", "size_complete": false}}}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "405": {"$ref": "#/components/responses/DiskDetailMethod"},
          "500": {"$ref": "#/components/responses/Internal"},
          "503": {"$ref": "#/components/responses/Unavailable"}
        }
      },
      "patch": {
        "summary": "Atomically edit disk metadata",
        "description": "Omitted fields remain unchanged. Null or empty notes/serial clear the field; label/capacity cannot be null. Empty requests, unknown fields, trailing JSON values and query parameters are rejected. Labels remain case-sensitive and unique. No inventory/content mutation. Body maximum 65536 bytes; Content-Type must be application/json.",
        "parameters": [{"$ref": "#/components/parameters/ID"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateDisk"}, "example": {"label": "renamed archive", "notes": null}}}},
        "responses": {
          "200": {"description": "Updated disk detail", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DiskDetail"}}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "405": {"$ref": "#/components/responses/DiskDetailMethod"},
          "409": {"$ref": "#/components/responses/DiskConflict"},
          "413": {"$ref": "#/components/responses/TooLarge"},
          "415": {"$ref": "#/components/responses/UnsupportedMedia"},
          "500": {"$ref": "#/components/responses/Internal"},
          "503": {"$ref": "#/components/responses/Unavailable"}
        }
      }
    },
    "/api/v1/disks/{id}/snapshots": {
      "get": {
        "summary": "Page through a disk's complete inventories",
        "description": "Ordered by captured_at descending, then snapshot ID descending. Older inventories imported later do not outrank newer captures. Existing disks without complete inventories return an empty page; nonexistent disks return 404. Keep disk ID and limit unchanged when following a cursor. Revision is the highest completed snapshot ID, or 0. A successful offline import on any disk invalidates cursors; unchanged restarts, disk creation, failed imports and removal of incomplete inventories preserve them. Cursors are for the issuing catalog. Unknown, empty and repeated scalar parameters are rejected.",
        "parameters": [
          {"$ref": "#/components/parameters/ID"},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 200, "default": 50}},
          {"name": "cursor", "in": "query", "schema": {"type": "string", "minLength": 1, "maxLength": 2048}, "description": "Opaque next_cursor from the preceding page"}
        ],
        "responses": {
          "200": {"description": "Complete snapshot page", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SnapshotPage"}, "examples": {
            "lastPage": {"value": {"disk_id": "3", "revision": "3", "items": [{"id": "3", "disk_id": "3", "state": "complete", "captured_at": "1970-01-01T00:00:10.000000123Z", "imported_at": "2026-10-05T12:00:00Z", "capture_provenance": "explicit", "input_path": "/inventories/disk-2.cshd", "input_format": "cshd", "file_count": "1", "content_count": "1"}], "next_cursor": null}},
            "emptyDisk": {"value": {"disk_id": "4", "revision": "3", "items": [], "next_cursor": null}}
          }}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "405": {"description": "Only GET and HEAD are allowed; Allow header is GET, HEAD", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "method_not_allowed", "message": "method not allowed"}}}}},
          "409": {"description": "Inventory changed since the cursor was issued; restart pagination", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "stale_cursor", "message": "cursor belongs to a different inventory revision"}}}}},
          "503": {"$ref": "#/components/responses/Unavailable"},
          "500": {"$ref": "#/components/responses/Internal"}
        }
      }
    },
    "/api/v1/snapshots/{id}": {
      "get": {
        "summary": "Read a complete inventory's provenance and counters",
        "parameters": [{"$ref": "#/components/parameters/ID"}],
        "responses": {
          "200": {"description": "Complete snapshot", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Snapshot"}}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "503": {"$ref": "#/components/responses/Unavailable"},
          "500": {"$ref": "#/components/responses/Internal"}
        }
      }
    }
  },
  "components": {
    "parameters": {
      "ID": {"name": "id", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/ID"}},
      "Scope": {"name": "scope", "in": "query", "schema": {"type": "string", "enum": ["current", "history"], "default": "current"}}
    },
    "responses": {
      "DiskCollectionMethod": {"description": "Allowed methods: GET, HEAD, POST", "headers": {"Allow": {"schema": {"type": "string", "const": "GET, HEAD, POST"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "DiskConflict": {"description": "Label already belongs to another disk", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "conflict", "message": "request conflicts with catalog state"}}}}},
      "StaleCursor": {"description": "Inventory changed; restart pagination", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "stale_cursor", "message": "cursor belongs to a different inventory revision"}}}}},
...
      "Content": {"description": "Content summary. Observation count always includes all complete historical observations. Unknown size is null, known zero is string 0.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Content"}, "example": {"id": "1", "hash": {"algorithm": "sha256", "hex": "ab"}, "size": "0", "scope": "current", "location_count": "4", "disk_count": "3", "observation_count": "4", "current_location_count": "4", "current_disk_count": "3"}}}},
      "Invalid": {"description": "Invalid or mismatched request/cursor", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "invalid_request", "message": "invalid request: limit must be between 1 and 200"}}}}},
      "NotFound": {"description": "Resource absent or visible only in an incomplete inventory", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "not_found", "message": "resource not found"}}}}},
      "Unavailable": {"description": "Catalog/session unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "catalog_unavailable", "message": "catalog is unavailable"}}}}},
      "Internal": {"description": "Unexpected server failure", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": {"code": "internal_error", "message": "internal server error"}}}}}
    },
    "schemas": {
      "CreateDisk": {
        "type": "object", "required": ["label", "capacity"], "additionalProperties": false,
        "properties": {
          "label": {"type": "string", "minLength": 1},
          "capacity": {"type": "string", "pattern": "^[0-9]+$", "description": "Decimal bytes, maximum 9223372036854775807"},
          "notes": {"type": ["string", "null"]}, "serial": {"type": ["string", "null"]}
        }
      },
...
      "DiskDetail": {
        "type": "object", "required": ["id", "label", "notes", "serial", "capacity", "latest_snapshot", "cataloged"],
        "properties": {
          "id": {"$ref": "#/components/schemas/ID"}, "label": {"type": "string"},
          "notes": {"type": ["string", "null"]}, "serial": {"type": ["string", "null"]},
          "capacity": {"$ref": "#/components/schemas/Decimal"},
          "latest_snapshot": {"$ref": "#/components/schemas/NullableSnapshot"},
          "cataloged": {"type": ["object", "null"], "required": ["file_count", "content_count", "known_bytes", "unknown_size_file_count", "size_complete"], "properties": {
            "file_count": {"$ref": "#/components/schemas/Count"}, "content_count": {"$ref": "#/components/schemas/Count"},
            "known_bytes": {"$ref": "#/components/schemas/Decimal"}, "unknown_size_file_count": {"$ref": "#/components/schemas/Count"},
            "size_complete": {"type": "boolean"}
          }}
        }
      },
      "DiskPage": {
        "type": "object", "required": ["revision", "items", "next_cursor"],
        "properties": {
          "revision": {"$ref": "#/components/schemas/Decimal"},
          "items": {"type": "array", "items": {"$ref": "#/components/schemas/DiskSummary"}},
          "next_cursor": {"type": ["string", "null"]}
        }
      },
...
      "Decimal": {"type": "string", "pattern": "^(0|[1-9][0-9]*)$", "description": "Nonnegative signed 64-bit integer as decimal text"},
      "ContentPage": {
        "type": "object", "required": ["scope", "revision", "filters", "items", "next_cursor"],
        "properties": {
          "scope": {"type": "string", "enum": ["current", "history"]},
          "revision": {"$ref": "#/components/schemas/Decimal"},
          "filters": {"type": "object", "required": ["disk_id", "directory", "replica_metric", "other_replicas", "min_other_replicas", "max_other_replicas"], "properties": {
            "disk_id": {"type": ["string", "null"], "pattern": "^[1-9][0-9]*$"},
            "directory": {"type": "string"},
            "replica_metric": {"type": "string", "enum": ["disks", "locations"]},
            "other_replicas": {"type": ["string", "null"], "pattern": "^[0-9]+$"},
            "min_other_replicas": {"type": ["string", "null"], "pattern": "^[0-9]+$"},
            "max_other_replicas": {"type": ["string", "null"], "pattern": "^[0-9]+$"}
          }},
          "items": {"type": "array", "items": {"$ref": "#/components/schemas/Content"}},
          "next_cursor": {"type": ["string", "null"]}
        }
      },
...
      "ID": {"type": "string", "pattern": "^[1-9][0-9]*$", "description": "Positive signed 64-bit ID as decimal text"},
      "Count": {"type": "string", "pattern": "^(0|[1-9][0-9]*)$"},
...
      "Algorithm": {"type": "string", "enum": ["md4", "md5", "sha1", "sha256", "sha384", "sha3_224", "sha3_256", "sha3_384", "sha3_512", "sha512"]},
...
      "ObservationSummary": {
        "type": "object", "required": ["observation", "snapshot", "disk", "content", "other_location_count", "other_disk_count"],
        "properties": {"observation": {"$ref": "#/components/schemas/Observation"}, "snapshot": {"$ref": "#/components/schemas/Snapshot"}, "disk": {"$ref": "#/components/schemas/Disk"}, "content": {"$ref": "#/components/schemas/Content"}, "other_location_count": {"$ref": "#/components/schemas/Count"}, "other_disk_count": {"$ref": "#/components/schemas/Count"}}
      },
...
      "ObservationPage": {
        "type": "object", "required": ["scope", "revision", "items", "next_cursor"],
        "properties": {"scope": {"type": "string", "enum": ["current", "history"]}, "revision": {"$ref": "#/components/schemas/Count"}, "items": {"type": "array", "items": {"$ref": "#/components/schemas/ContentObservation"}}, "next_cursor": {"type": ["string", "null"]}}
      },
      "Error": {
        "type": "object", "required": ["error"],
        "properties": {"error": {"type": "object", "required": ["code", "message"], "properties": {"code": {"type": "string", "enum": ["invalid_request", "not_found", "stale_cursor", "conflict", "catalog_unavailable", "internal_error", "method_not_allowed", "request_too_large", "unsupported_media_type"]}, "message": {"type": "string"}}}}
      }
    }
  }
}